==> 6_RB_gaussian_example_figure.R <==
# Example rangebagging vs gaussian figure


# Load libraries

library(BIEN)
library(ggplot2)
library(S4DM)
library(sf)
library(terra)
library(tidyterra)

# Make a temporary directory to store climate data

temp <- tempdir()

# Get some occurrence data

#tv <- BIEN_occurrence_species(species = "Trillium vaseyi")
data("sample_points")


# Get environmental data
# To make things a bit faster and easier, we'll limit ourselves to the 2 variables (mean temperature and annual precipitation)


# env <- worldclim_global(var = "bio",
#                          res = 10,
#                          path = temp)
# env <- env[[c(1,12)]]

env <- rast(system.file('ex/sample_env.tif', package="S4DM"))  

env <- env[[1]]

# And we'll rescale the variables as well

env <- scale(env)

# Just to take a look to make sure we didn't mess anything up

plot(env)


# make models

sample_points_bg <- get_env_bg(coords = sample_points[c("longitude","latitude")],
                               env = env,
                               width = 50000) #note that we used a small set of background points to expedite model fitting

sample_points_pres <- get_env_pres(coords = sample_points[c("longitude","latitude")],
                                   env = env)


rbna <- S4DM::fit_plug_and_play(presence = sample_points_pres$env,
                                background = sample_points_bg$env,
                                presence_method = "rangebagging",
                                background_method = "none",
                                d=1)

naga <- S4DM::fit_plug_and_play(presence = sample_points_pres$env,
                                background = sample_points_bg$env,
                                presence_method = "none",
                                background_method = "gaussian")



# make data for plotting


pred_data <- data.frame(wc2.1_10m_bio_1 = seq(from=-3,to=3,by=.01))

rbna_pred <- S4DM::project_plug_and_play(pnp_model = rbna,
                                         data = pred_data)

naga_pred <- S4DM::project_plug_and_play(pnp_model = naga,
                                         data = pred_data)


rbga_pred <- rbna_pred/naga_pred


library(ggplot2)
library(tidyverse)

plot_data <-
  bind_rows(
    data.frame(model = "rangebagging",
               env = pred_data,
               value= rbna_pred),
    
    data.frame(model = "gaussian",
               env = pred_data,
               value = naga_pred),
    
    data.frame(model = "rangebagging/gaussian",
               env = pred_data,
               value = rbga_pred)
    
  )%>%
  rename(env = wc2.1_10m_bio_1)

ggplot(data = plot_data,
       mapping = aes(x=env,
                     y=value,
                     color=model),
       alpha=0.5)+
  geom_line()


##############################

# simpler model to show the overall idea

plot_data <-
bind_rows(
  data.frame(env = seq(-3,3,.1),
             model = "Background",
             value= dnorm(mean = 0,sd = 1,
                   x = seq(-3,3,.1))),
  data.frame(env = seq(-3,3,.1),
             model = "Presence",
             value= dunif(min = -1,max = 1,
                   x = seq(-3,3,.1)))
  )%>%
  mutate(value = case_when(model=="Presence" & value > 0 ~1,.default = value))%>%
  pivot_wider(names_from = model,values_from = value) %>%
  mutate("Relative\nOccurrence\nRate" = Presence/Background)%>%
  pivot_longer(cols = 2:4,names_to = "model")

example_plot <-
plot_data %>%
ggplot(mapping = aes(x=env,y=value,color=model))+
  geom_line(linewidth=1,alpha=0.75)+
  theme_bw()+
  scale_color_manual(values = c('pink',"grey","purple"))+
  xlab("environment")+
  ylab("f(environment)")+
  theme(legend.title = element_blank())

ggsave(filename = "figures/example_rb_gaussian.jpeg",
       plot = example_plot,
       width = 5,height = 5,units = "in",dpi = 600)


ggsave(filename = "figures/example_rb_gaussian.png",
       plot = example_plot,
       width = 5,height = 5,units = "in",dpi = 600)

######################################################

# Figure for revision with different model types


rbna <- S4DM::fit_plug_and_play(presence = sample_points_pres$env,
                                background = sample_points_bg$env,
                                presence_method = "rangebagging",
                                background_method = "none",
                                d=1)

kdena <- S4DM::fit_plug_and_play(presence = sample_points_pres$env,
                                background = sample_points_bg$env,
                                presence_method = "kde",
                                background_method = "none",
                                d=1)

gana <- S4DM::fit_plug_and_play(presence = sample_points_pres$env,
                                background = sample_points_bg$env,
                                presence_method = "gaussian",
                                background_method = "none")

vina <- S4DM::fit_plug_and_play(presence = sample_points_pres$env,
                                background = sample_points_bg$env,
                                presence_method = "vine",
                                background_method = "none")

lbna <- S4DM::fit_plug_and_play(presence = sample_points_pres$env,
                                background = sample_points_bg$env,
                                presence_method = "lobagoc",
                                background_method = "none")



rbna_pred <- S4DM::project_plug_and_play(pnp_model = rbna,
                                         data = pred_data)

gana_pred <- S4DM::project_plug_and_play(pnp_model = gana,
                                         data = pred_data)


kdena_pred <- S4DM::project_plug_and_play(pnp_model = kdena,
                                         data = pred_data)


lbna_pred <- S4DM::project_plug_and_play(pnp_model = lbna,
                                          data = pred_data)

vina_pred <- S4DM::project_plug_and_play(pnp_model = vina,
                                          data = pred_data)


dist_type_plot_data <-
  bind_rows(
    data.frame(model = "rangebagging",
               env = pred_data,
               value= rbna_pred),
    
    data.frame(model = "gaussian",
               env = pred_data,
               value = gana_pred),
    
    data.frame(model = "kde",
               env = pred_data,
               value = kdena_pred),
    
    data.frame(model = "vine",
               env = pred_data,
               value= vina_pred),
    
    data.frame(model = "lobagoc",
               env = pred_data,
               value= lbna_pred)
    
    
    
  )%>%
  rename(env = wc2.1_10m_bio_1)

dist_type_plot_data <- dist_type_plot_data %>%
  group_by(model)%>%
  mutate(total_pred = sum(value))%>%
  mutate(ror = value/total_pred)


dist_type_plot <-
dist_type_plot_data %>%
  ggplot(mapping = aes(x=env,y = ror))+
  geom_line()+
  facet_wrap(~model,ncol=1)+
  theme_bw()+
  ylab("Relative Occurrence Rate")+
  xlab("Environmental Value")+
  scale_x_continuous(expand = c(0,0))

ggsave(plot = dist_type_plot,
       filename = "figures/algorithm_example_plot.jpg",
       height = 6,width = 4,units = "in",dpi = 300)


##############################################################
env2d <- rast(system.file('ex/sample_env.tif', package="S4DM"))[[1:2]] %>% scale()


# And we'll rescale the variables as well
plot(env2d)

sample_points_bg2d <- get_env_bg(coords = sample_points[c("longitude","latitude")],
                               env = env2d,
                               width = 50000) #note that we used a small set of background points to expedite model fitting

sample_points_pres2d <- get_env_pres(coords = sample_points[c("longitude","latitude")],
                                   env = env2d)


rbna2d <- S4DM::fit_plug_and_play(presence = sample_points_pres2d$env,
                                background = sample_points_bg2d$env,
                                presence_method = "rangebagging",
                                background_method = "none",
                                d=2)

lbna2d <- S4DM::fit_plug_and_play(presence = sample_points_pres2d$env,
                                  background = sample_points_bg2d$env,
                                  presence_method = "lobagoc",
                                  background_method = "none")

gana2d <- S4DM::fit_plug_and_play(presence = sample_points_pres2d$env,
                                  background = sample_points_bg2d$env,
                                  presence_method = "gaussian",
                                  background_method = "none")

kdna2d <- S4DM::fit_plug_and_play(presence = sample_points_pres2d$env,
                                  background = sample_points_bg2d$env,
                                  presence_method = "kde",
                                  background_method = "none")


vina2d <- S4DM::fit_plug_and_play(presence = sample_points_pres2d$env,
                                  background = sample_points_bg2d$env,
                                  presence_method = "vine",
                                  background_method = "none")
# 
# mn2d <- S4DM::fit_density_ratio(presence = sample_points_pres2d$env,
#                                   background = sample_points_bg2d$env,
#                                 method = "maxnet")
# 
# ul2d <- S4DM::fit_density_ratio(presence = sample_points_pres2d$env,
#                                 background = sample_points_bg2d$env,
#                                 method = "ulsif")
# 
# ru2d <- S4DM::fit_density_ratio(presence = sample_points_pres2d$env,
#                                 background = sample_points_bg2d$env,
#                                 method = "rulsif")


# Make pred data

  pred_data_temp <- data.frame(temp = seq(from=-4,to=4,by=.01))
  pred_data_precip <- data.frame(precip = seq(from=-4,to=4,by=.01))
  combined_preds<- cross_join(pred_data_temp,pred_data_precip)
  names(combined_preds) <- names(env2d)

# Do predictions  
    
  rbna_pred_2d <- S4DM::project_plug_and_play(pnp_model = rbna2d,
                                           data =combined_preds)
  
  lbna_pred_2d <- S4DM::project_plug_and_play(pnp_model = lbna2d,
                                              data = combined_preds)
  
  gana_pred_2d <- S4DM::project_plug_and_play(pnp_model = gana2d,
                                              data = combined_preds)
  
  kdna_pred_2d <- S4DM::project_plug_and_play(pnp_model = kdna2d,
                                              data = combined_preds)
  
  vina_pred_2d <- S4DM::project_plug_and_play(pnp_model = vina2d,
                                              data = combined_preds)
  
  # mn_pred_2d <- S4DM::project_density_ratio(dr_model = mn2d,
  #                                             data = combined_preds)
  # 
  # ru_pred_2d <- S4DM::project_density_ratio(dr_model = ru2d,
  #                                           data = combined_preds)
  # 
  # ul_pred_2d <- S4DM::project_density_ratio(dr_model = ul2d,
  #                                           data = combined_preds)
  


combined_preds_data <-
combined_preds %>%
  bind_cols(data.frame(model="rangebagging",
                       value = rbna_pred_2d)) %>%

  bind_rows(combined_preds %>%
              bind_cols(data.frame(model="lobagoc",
                                   value = lbna_pred_2d))) %>%
  bind_rows(combined_preds %>%
              bind_cols(data.frame(model="gaussian",
                                   value = gana_pred_2d))) %>%

  bind_rows(combined_preds %>%
              bind_cols(data.frame(model="kde",
                                   value = kdna_pred_2d))) %>%
  
  bind_rows(combined_preds %>%
              bind_cols(data.frame(model="vine",
                                   value = vina_pred_2d))) %>%
  # bind_rows(combined_preds %>%
  #             bind_cols(data.frame(model="maxnet",
  #                                  value = mn_pred_2d))) %>%
  # 
  # bind_rows(combined_preds %>%
  #             bind_cols(data.frame(model="rulsif",
  #                                  value = ru_pred_2d))) %>%
  # 
  # bind_rows(combined_preds %>%
  #             bind_cols(data.frame(model="ulsif",
  #                                  value = ul_pred_2d))) %>%
  
  group_by(model) %>%
  mutate(model_total = sum(value)) %>%
  mutate(ror = value / model_total) 

combined_preds_plot <-  
combined_preds_data %>%
  mutate(model = factor(model,c("gaussian","kde","vine","rangebagging","lobagoc"))) %>%
  ggplot(mapping = aes(x=wc2.1_10m_bio_1,
                       y = wc2.1_10m_bio_12,
                       fill = ror))+
  geom_raster()+
  facet_wrap(~model)+
  scale_x_continuous(expand = c(0,0),limits = c(-3,3))+
  scale_y_continuous(expand = c(0,0),limits=c(-1,4))+
  scale_fill_gradient(low = "white",
                      high = "blue", 
                      na.value = 0,
                      breaks = range(combined_preds_data$ror),
                      labels = c("Low", "High"))+
  theme_bw()+
  labs(fill = "Relative\nOccurrence\nRate")+
  xlab("Temperature")+
  ylab("Precipitation")+
  geom_point(data = sample_points_pres2d$env,
             mapping = aes(x=wc2.1_10m_bio_1,
                           y = wc2.1_10m_bio_12),
             inherit.aes = FALSE,
             size = .1)

ggsave(plot = combined_preds_plot,
       filename = "figures/algoritm_example_plot_2d.jpg",
       width = 5,
       height = 5,
       units = "in",
       dpi = 600)
